+++ openmp/priority_queue.h
@@ -9,6 +9,6 @@ void push(heap_t *heap, struct node_t *node);
 
 struct node_t* pop(heap_t* heap);
 
-void insert_values_in_queue(unsigned long long int* frequecy, heap_t* h);
+void insert_values_in_queue(int* frequecy, heap_t* h);
 
 #endif /* PRIORITY_QUEUE_H */
